Update half_blast.py
This version will take a multiple fasta as an input. It will cut the protein in half, create a database for blastp based on the first half of the protein and blast the second half against it. It will then save any hits with evalue < 100 to half_blast.out. For my application evalues of <25 seemed to be biologically relevant, but that is likely dependent on the system.